wip check filename attribute
