+++ src/integrationTest/kotlin/io/github/endzeitbegins/nifi/flowovertcp/IntegrationTest.kt
@@ -4,23 +4,27 @@ import io.github.endzeitbegins.nifi.flowovertcp.gateways.KtorNiFiApiGateway
 import io.github.endzeitbegins.nifi.flowovertcp.gateways.NiFiApiGateway
 import io.github.endzeitbegins.nifi.flowovertcp.gateways.setUpNiFiFlow
 import io.github.endzeitbegins.nifi.flowovertcp.testcontainers.NiFiContainerProvider
+import io.github.endzeitbegins.nifi.flowovertcp.testing.assertions.contentAsJson
+import io.github.endzeitbegins.nifi.flowovertcp.testing.assertions.sha256HashSum
+import io.github.endzeitbegins.nifi.flowovertcp.testing.waitFor
+import io.github.endzeitbegins.nifi.flowovertcp.utils.deleteRegularFilesRecursively
+import io.github.endzeitbegins.nifi.flowovertcp.utils.sha256
+import kotlinx.serialization.json.Json
+import kotlinx.serialization.json.decodeFromStream
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestInstance
 import strikt.api.expect
+import strikt.assertions.hasEntry
 import strikt.assertions.isEqualTo
 import strikt.java.allBytes
 import strikt.java.exists
 import strikt.java.size
-import java.io.IOException
 import java.nio.file.*
-import java.nio.file.attribute.BasicFileAttributes
-import java.security.MessageDigest
 import java.time.Duration
-import java.time.Instant
 import java.util.*
-import java.util.concurrent.TimeoutException
 import kotlin.io.path.*
 import kotlin.random.Random
+
 
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 class IntegrationTest {
@@ -71,59 +75,15 @@ class IntegrationTest {
                 that(contentFilePath)
                     .exists()
                     .and { size.isEqualTo(testFile.fileSize) }
-                    .and { allBytes().get(ByteArray::sha256).isEqualTo(testFile.fileSha256Hash) }
+                    .and { allBytes().sha256HashSum.isEqualTo(testFile.fileSha256Hash) }
 
                 that(attributesFilePath)
                     .exists()
-                // todo test attributes exist ...
+                    .contentAsJson()
+                    .hasEntry("filename", testFile.fileName)
             }
         }
     }
-
-
-}
-
-private fun Path.deleteRegularFilesRecursively() {
-    Files.walkFileTree(this, object : FileVisitor<Path> {
-        override fun preVisitDirectory(dir: Path, attrs: BasicFileAttributes): FileVisitResult {
-            return FileVisitResult.CONTINUE
-        }
-
-        override fun visitFile(file: Path, attrs: BasicFileAttributes): FileVisitResult {
-            if (!Files.isHidden(file)) {
-                file.deleteExisting()
-            }
-
-            return FileVisitResult.CONTINUE
-        }
-
-        override fun visitFileFailed(file: Path, exc: IOException?): FileVisitResult {
-            if (exc != null) {
-                throw exc
-            }
-
-            return FileVisitResult.CONTINUE
-        }
-
-        override fun postVisitDirectory(dir: Path?, exc: IOException?): FileVisitResult {
-            if (exc != null) {
-                throw exc
-            }
-
-            return FileVisitResult.CONTINUE
-        }
-
-    })
-//    @OptIn(ExperimentalPathApi::class)
-//    this.visitFileTree {
-//        onVisitFile { file, _ ->
-//            if (!file.isHidden()) {
-//                file.deleteExisting()
-//            }
-//
-//            FileVisitResult.CONTINUE
-//        }
-//    }
 }
 
 private data class TestFile(
@@ -131,33 +91,3 @@ private data class TestFile(
     val fileSize: Long,
     val fileSha256Hash: String,
 )
-
-private val ByteArray.sha256: String
-    get() {
-        val md = MessageDigest.getInstance("SHA-256")
-        val digest = md.digest(this)
-
-        return digest.joinToString(separator = "") { byte -> "%02x".format(byte) }
-    }
-
-
-private fun waitFor(
-    duration: Duration,
-    errorMessage: String = "Failed to meet predicate after $duration.",
-    predicate: () -> Boolean,
-) {
-    val startTime = Instant.now()
-    val timeout = startTime.plus(duration)
-
-    do {
-        if (predicate()) {
-            return
-        }
-
-        Thread.sleep(250)
-    } while (Instant.now() < timeout)
-
-    throw TimeoutException(errorMessage)
-}
-
-
